requirements.txt
asgiref==3.4.1
Django==3.2.11
future==0.18.2
pkg-resources==0.0.0
psycopg2==2.9.3
python-dateutil==2.8.2
python-whois==0.7.3
python-whois-extended==0.7.12
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.0.1
